refact: 제목, 가격을 SessionInfo 클래스로 분리

+++ src/main/java/nextstep/courses/domain/session/PaidSession.java
@@ -5,7 +5,7 @@ import nextstep.payments.domain.Payment;
 import java.time.LocalDateTime;
 
 public class PaidSession extends Session {
-    public PaidSession(final String title, final double price, final LocalDateTime startDate, final LocalDateTime endDate) {
+    public PaidSession(final String title, final long price, final LocalDateTime startDate, final LocalDateTime endDate) {
         super(title, price, startDate, endDate);
     }
